+++ ui/app/src/components/PathNavigator/PathNavigator.tsx
@@ -56,6 +56,8 @@ import { GetChildrenOptions, PartialSxRecord } from '../../models';
 import { pushDialog } from '../../state/actions/dialogStack';
 import { nanoid } from 'nanoid';
 import { showItemMegaMenu } from '../../state/actions/dialogs';
+import usePossibleTranslation from '../../hooks/usePossibleTranslation';
+import TranslationOrText from '../../models/TranslationOrText';
 
 interface Menu {
 	path?: string;
@@ -72,7 +74,7 @@ export type PathNavigatorClassKey = 'root' | 'body' | 'searchRoot';
 
 export interface PathNavigatorProps {
 	id: string;
-	label: string;
+	label: TranslationOrText;
 	rootPath: string;
 	sortStrategy?: GetChildrenOptions['sortStrategy'];
 	order?: GetChildrenOptions['order'];
@@ -126,15 +128,15 @@ export interface PathNavigatorStateProps {
 // };
 
 export function PathNavigator(props: PathNavigatorProps) {
+	const title = usePossibleTranslation(props.label ?? '(No name)');
 	// region const { ... } = props;
 	const {
-		label = '(No name)',
 		icon,
 		expandedIcon,
 		collapsedIcon,
 		container,
 		rootPath: path,
-		id = label.replace(/\s/g, ''),
+		id = title.replace(/\s/g, ''),
 		limit = 10,
 		locale,
 		excludes,
@@ -412,7 +414,7 @@ export function PathNavigator(props: PathNavigatorProps) {
 				itemsByPath={itemsByPath}
 				icon={expandedIcon && collapsedIcon ? (state.collapsed ? collapsedIcon : expandedIcon) : icon}
 				container={container}
-				title={label}
+				title={title}
 				onChangeCollapsed={onChangeCollapsed}
 				onHeaderButtonClick={state.collapsed ? void 0 : onHeaderButtonClick}
 				onCurrentParentMenu={onCurrentParentMenu}