+++ ui/app/src/CHANGELOG.md
@@ -55,6 +55,7 @@
   * Added dialogs: `CancelPackageDialog`, `BulkCancelPackageDialog`, `PublishPackageReviewDialog`, `PublishingPackageResubmitDialog`, `ViewPackagesDialog`.
   * [RenameAssetDialog] Removed `path` and `value` properties. Replaced with `item` property.
   * [ItemDisplay] Updated `item` prop to be of type `LightItem | ContentItem`.
+  * `PathNavigator/PathNavigatorTree`: Updated `label` prop to be of type `TranslationOrText`.
 * [hooks]
   * Removed `useLogicResource` hook.
   * Removed `useSelectorResource` hook.